feat(kernel): replace event bus middleware with http kernel
