+++ src/Config/types.ts
@@ -1,6 +1,6 @@
-import { type EventSubscriber } from '@/Event';
 import { type HttpMiddleware } from '@/Http';
 import { type StaticFilesOptions } from '@/Http/Files';
+import { type EventSubscriber } from '@/Kernel';
 
 export type Controller = new (...args: unknown[]) => unknown;
