+++ src/Event/index.ts
@@ -1,2 +1,5 @@
-export * from './types';
-export * from './EventBus';
+/**
+ * @deprecated use `httpKernel` from `@/Kernel/HttpKernel` instead
+ * {@see httpKernel}
+ */
+export { httpKernel as eventBusMiddleware } from '@/Kernel/HttpKernel';
